Use uint for message counts.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@3407 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/globals.h
@@ -67,8 +67,8 @@ extern	const char *const version;
 extern	const char *const system_config_file;
 
 /* for msgcounts.c */
-extern	long	 msgs_good;
-extern 	long	 msgs_bad;
+extern	uint		 msgs_good;
+extern	uint		 msgs_bad;
 
 /* from html.h */
 extern	bool strict_check;
